+++ Server/WebGame/PlayerSession.h
@@ -21,11 +21,11 @@ public:
 
 	virtual void		OnConnect(void);
 	virtual void		OnClose(void);
-	virtual void		OnError(UINT32 dwErrorNo);
-	virtual void		OnRecv(const char* pBuf, UINT32 dwLen);
+	virtual void		OnError(unsigned int dwErrorNo);
+	virtual void		OnRecv(const char* pBuf, unsigned int dwLen);
 	virtual void		Release(void);
 	virtual char*		GetRecvBuf() { return m_dataRecvBuf; }
-	virtual UINT32		GetRecvSize() { return 64 * 1024; };
+	virtual unsigned int		GetRecvSize() { return 64 * 1024; };
 
 	bool				OnPlayerRequestGameTest(CPlayerSession& refSession, google::protobuf::Message& refMsg);
 	bool				OnPlayerRequestGameEnter(CPlayerSession& refSession, google::protobuf::Message& refMsg);
@@ -37,7 +37,7 @@ private:
 	CallBackDispatcher::ProtoCallBackDispatch<CPlayerSession, CPlayerSession> m_oProtoDispatch;
 
 protected:
-	UINT64 m_qwPlayerId;
+	unsigned long long m_qwPlayerId;
 };
 
 class CWebSocketPlayerSession : public CPlayerSession
